+++ src/Resources/GameContentVersionResource/Pages/EditGameContentVersion.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\BrowserGame\GameCoreFilament\Resources\GameContentVersionResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use Liberu\BrowserGame\GameCoreFilament\Resources\GameContentVersionResource;
+
+final class EditGameContentVersion extends EditRecord
+{
+    protected static string $resource = GameContentVersionResource::class;
+}