+++ src/Resources/GameRulesetResource/Pages/EditGameRuleset.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\BrowserGame\GameCoreFilament\Resources\GameRulesetResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use Liberu\BrowserGame\GameCoreFilament\Resources\GameRulesetResource;
+
+final class EditGameRuleset extends EditRecord
+{
+    protected static string $resource = GameRulesetResource::class;
+}