+++ src/Resources/GameWorldResource/Pages/EditGameWorld.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\BrowserGame\GameCoreFilament\Resources\GameWorldResource\Pages;
+
+use Filament\Resources\Pages\EditRecord;
+use Liberu\BrowserGame\GameCoreFilament\Resources\GameWorldResource;
+
+final class EditGameWorld extends EditRecord
+{
+    protected static string $resource = GameWorldResource::class;
+}