+++ src/Resources/GameWorldResource/Pages/ListGameWorlds.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liberu\BrowserGame\GameCoreFilament\Resources\GameWorldResource\Pages;
+
+use Filament\Resources\Pages\ListRecords;
+use Liberu\BrowserGame\GameCoreFilament\Resources\GameWorldResource;
+
+final class ListGameWorlds extends ListRecords
+{
+    protected static string $resource = GameWorldResource::class;
+}